Various scheduling cleanup
