+++ src/task/mod.rs
@@ -1,7 +1,7 @@
 pub mod process;
 pub mod scheduler;
 
-pub use self::process::{Process, ProcessId, ProcessRef, ProcessTable, State};
+pub use self::process::{Process, ProcessId, ProcessRef, State};
 pub use self::scheduler::{global_sched, Scheduling};
 
 const MAX_PID: usize = <usize>::max_value() - 1;
